requirements.txt
django==2.1.5
djangorestframework==3.9.1
psycopg2==2.7.7
